requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# requirements.txt — Laboratório 07: Especialização de LLMs com LoRA e QLoRA
# ══════════════════════════════════════════════════════════════════════════════
# Instalação:
#   pip install -r requirements.txt
#
# Nota: bitsandbytes requer CUDA. Em ambiente CPU, remova ou comente a linha.
# Testado com Python 3.10+ e CUDA 12.x
# ══════════════════════════════════════════════════════════════════════════════

# ── Geração de Dataset ────────────────────────────────────────────────────────
openai>=1.30.0            # Cliente oficial (compatível com OpenRouter)
python-dotenv>=1.0.0      # Carregamento de variáveis de ambiente (.env)

# ── Modelo e Tokenizer ────────────────────────────────────────────────────────
transformers>=4.40.0      # AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
datasets>=2.19.0          # load_dataset (formato .jsonl)
accelerate>=0.29.0        # device_map="auto", mixed precision

# ── Quantização 4-bit (QLoRA) ─────────────────────────────────────────────────
bitsandbytes>=0.43.0      # BitsAndBytesConfig: NF4, compute_dtype

# ── PEFT / LoRA ───────────────────────────────────────────────────────────────
peft>=0.10.0              # LoraConfig, get_peft_model, prepare_model_for_kbit_training

# ── Treinamento Supervisionado ────────────────────────────────────────────────
trl>=0.8.6                # SFTTrainer (Supervised Fine-Tuning Trainer)

# ── Monitoramento (opcional) ──────────────────────────────────────────────────
tensorboard>=2.16.0       # Visualização de métricas (acessar via: tensorboard --logdir outputs/)
